azure-form-filler/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
azure-ai-documentintelligence==1.0.0b2
pypdf==4.3.1
reportlab==4.2.2
Pillow==10.4.0
